Финальная работа/requirements.txt
progressbar2>=3.38.0
requests>=2.18.4
